requirements.txt
django
django-crispy-forms
django-filter
djangorestframework
requests
